common: remove unnecessary casts (#106)

+++ _studio/mfx_lib/shared/src/libmfxsw_decode.cpp
@@ -483,9 +483,9 @@ mfxStatus MFXVideoDECODE_DecodeFrameAsync(mfxSession session, mfxBitstream *bs, 
 
         memset(&task, 0, sizeof(MFX_TASK));
         mfxRes = session->m_pDECODE->DecodeFrameCheck(bs, surface_work, surface_out, &task.entryPoint);
-        MFX_CHECK(mfxRes >= 0 || MFX_ERR_MORE_DATA_SUBMIT_TASK == static_cast<int>(mfxRes)
-                  || MFX_ERR_MORE_DATA == static_cast<int>(mfxRes)
-                  || MFX_ERR_MORE_SURFACE == static_cast<int>(mfxRes), mfxRes);
+        MFX_CHECK(mfxRes >= 0 || MFX_ERR_MORE_DATA_SUBMIT_TASK == mfxRes
+                              || MFX_ERR_MORE_DATA             == mfxRes
+                              || MFX_ERR_MORE_SURFACE          == mfxRes, mfxRes);
 
         // source data is OK, go forward
         if (task.entryPoint.pRoutine)
@@ -514,7 +514,7 @@ mfxStatus MFXVideoDECODE_DecodeFrameAsync(mfxSession session, mfxBitstream *bs, 
             }
         }
 
-        if (MFX_ERR_MORE_DATA_SUBMIT_TASK == static_cast<int>(mfxRes))
+        if (MFX_ERR_MORE_DATA_SUBMIT_TASK == mfxRes)
         {
             mfxRes = MFX_WRN_DEVICE_BUSY;
         }
